Wybor akcji marcin

+++ Pokemony v1.2/Pokemony v1.2/Atak.cpp
@@ -2,7 +2,7 @@
 
 
 
-int Atak::getSila()
+float Atak::getSila()
 {
 	return Sila;
 }
